+++ MatrixKit/Controllers/MXKRecentListViewController.m
@@ -440,11 +440,14 @@
                     id<MXKRecentCellDataStoring> recentCellData = (id<MXKRecentCellDataStoring>)cellData;
                     if (recentCellData.isSuggestedRoom)
                     {
-                        [_delegate recentListViewController:self didSelectSuggestedRoom:recentCellData.spaceChildInfo];
+                        [_delegate recentListViewController:self
+                                     didSelectSuggestedRoom:recentCellData.roomSummary.spaceChildInfo];
                     }
                     else
                     {
-                        [_delegate recentListViewController:self didSelectRoom:recentCellData.roomSummary.roomId inMatrixSession:recentCellData.roomSummary.mxSession];
+                        [_delegate recentListViewController:self
+                                              didSelectRoom:recentCellData.roomIdentifier
+                                            inMatrixSession:recentCellData.mxSession];
                     }
                 }
             }