+++ MatrixKit/Models/RoomList/MXKSessionRecentsDataSource.m
@@ -51,7 +51,15 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
      */
     NSMutableDictionary<NSString*, NSArray<MXSpaceChildInfo *> *> *lastSuggestedRooms;
     
+    /**
+     Event listener of the current space used to update the UI if an event occurs.
+     */
     id spaceEventsListener;
+    
+    /**
+     Observer used to reload data when the space service is initialised
+     */
+    id spaceServiceDidInitialiseObserver;
 }
 
 /**
@@ -79,6 +87,8 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
         [self registerCellDataClass:MXKRecentCellData.class forCellIdentifier:kMXKRecentCellIdentifier];
         
         roomSummaryChangeThrottler = [[MXThrottler alloc] initWithMinimumDelay:roomSummaryChangeThrottlerDelay];
+        
+        [[MXKAppSettings standardAppSettings] addObserver:self forKeyPath:@"showAllRoomsInHomeSpace" options:0 context:nil];
     }
     return self;
 }
@@ -91,6 +101,10 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
     [[NSNotificationCenter defaultCenter] removeObserver:self name:kMXSessionDidLeaveRoomNotification object:nil];
     [[NSNotificationCenter defaultCenter] removeObserver:self name:kMXSessionDirectRoomsDidChangeNotification object:nil];
     
+    if (spaceServiceDidInitialiseObserver) {
+        [[NSNotificationCenter defaultCenter] removeObserver:spaceServiceDidInitialiseObserver];
+    }
+    
     [roomSummaryChangeThrottler cancelAll];
     roomSummaryChangeThrottler = nil;
     
@@ -101,6 +115,8 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
     
     searchPatternsList = nil;
     
+    [[MXKAppSettings standardAppSettings] removeObserver:self forKeyPath:@"showAllRoomsInHomeSpace" context:nil];
+
     [super destroy];
 }
 
@@ -196,16 +212,6 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
     return NO;
 }
 
-- (void)markAllAsRead
-{
-    // Clear unread count on all recent cells
-    for (NSUInteger i = 0; i < self.numberOfCells; i++)
-    {
-        id<MXKRecentCellDataStoring> cellData = [self cellDataAtIndex:i];
-        [cellData markAllAsRead];
-    }
-}
-
 - (void)searchWithPatterns:(NSArray*)patternsList
 {
     if (patternsList.count)
@@ -272,8 +278,11 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
     return 0;
 }
 
-
 #pragma mark - Events processing
+
+/**
+ Filtering in this method won't have any effect anymore. This class is not maintained.
+ */
 - (void)loadData
 {
     [[NSNotificationCenter defaultCenter] removeObserver:self name:kMXRoomSummaryDidChangeNotification object:nil];
@@ -281,6 +290,14 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
     [[NSNotificationCenter defaultCenter] removeObserver:self name:kMXSessionNewRoomNotification object:nil];
     [[NSNotificationCenter defaultCenter] removeObserver:self name:kMXSessionDidLeaveRoomNotification object:nil];
     [[NSNotificationCenter defaultCenter] removeObserver:self name:kMXSessionDirectRoomsDidChangeNotification object:nil];
+    
+    if (!self.mxSession.spaceService.isInitialised && !spaceServiceDidInitialiseObserver) {
+        MXWeakify(self);
+        spaceServiceDidInitialiseObserver = [[NSNotificationCenter defaultCenter] addObserverForName:MXSpaceService.didInitialise object:self.mxSession.spaceService queue:nil usingBlock:^(NSNotification * _Nonnull note) {
+            MXStrongifyAndReturnIfNil(self);
+            [self loadData];
+        }];
+    }
     
     // Reset the table
     [internalCellDataArray removeAllObjects];
@@ -297,20 +314,19 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
         if (!roomSummary.isConferenceUserRoom // @TODO Abstract this condition with roomSummary.hiddenFromUser
             && !roomSummary.hiddenFromUser)
         {
-            id<MXKRecentCellDataStoring> cellData = [[class alloc] initWithRoomSummary:roomSummary andRecentListDataSource:self];
+            id<MXKRecentCellDataStoring> cellData = [[class alloc] initWithRoomSummary:roomSummary dataSource:self];
             if (cellData)
             {
-                if (self.currentSpace == nil || [self.mxSession.spaceService isRoomWithId:roomSummary.roomId descendantOf:self.currentSpace.spaceId])
-                {
-                    [internalCellDataArray addObject:cellData];
-                }
+                [internalCellDataArray addObject:cellData];
             }
         }
     }
     
     for (MXSpaceChildInfo *childInfo in _suggestedRooms)
     {
-        id<MXKRecentCellDataStoring> cellData = [[class alloc] initWithSpaceChildInfo:childInfo andRecentListDataSource:self];
+        id<MXRoomSummaryProtocol> summary = [[MXRoomSummary alloc] initWithSpaceChildInfo:childInfo];
+        id<MXKRecentCellDataStoring> cellData = [[class alloc] initWithRoomSummary:summary
+                                                                        dataSource:self];
         if (cellData)
         {
             [internalCellDataArray addObject:cellData];
@@ -380,7 +396,7 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
             {
                 // Create a new instance to not modify the content of 'cellDataArray' (the copy is not a deep copy).
                 Class class = [self cellDataClassForCellIdentifier:kMXKRecentCellIdentifier];
-                id<MXKRecentCellDataStoring> cellData = [[class alloc] initWithRoomSummary:roomSummary andRecentListDataSource:self];
+                id<MXKRecentCellDataStoring> cellData = [[class alloc] initWithRoomSummary:roomSummary dataSource:self];
                 if (cellData)
                 {
                     [internalCellDataArray replaceObjectAtIndex:index withObject:cellData];
@@ -422,7 +438,7 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
             Class class = [self cellDataClassForCellIdentifier:kMXKRecentCellIdentifier];
 
             MXRoomSummary *roomSummary = [mxSession roomSummaryWithRoomId:roomId];
-            id<MXKRecentCellDataStoring> cellData = [[class alloc] initWithRoomSummary:roomSummary andRecentListDataSource:self];
+            id<MXKRecentCellDataStoring> cellData = [[class alloc] initWithRoomSummary:roomSummary dataSource:self];
             if (cellData)
             {
                 [internalCellDataArray addObject:cellData];
@@ -512,6 +528,22 @@ static NSTimeInterval const roomSummaryChangeThrottlerDelay = .5;
         }
     }
     return theRoomData;
+}
+
+#pragma mark - KVO
+
+- (void)observeValueForKeyPath:(NSString *)keyPath
+                      ofObject:(id)object
+                        change:(NSDictionary *)change
+                       context:(void *)context
+{
+    if (object == [MXKAppSettings standardAppSettings] && [keyPath isEqualToString:@"showAllRoomsInHomeSpace"])
+    {
+        if (self.currentSpace == nil)
+        {
+            [self loadData];
+        }
+    }
 }
 
 @end